requirements.txt
requests==2.32.3
pushbullet.py==0.12.0
bs4==0.0.2